requirements.txt
numpy~=2.1.2
gradio~=5.4.0
tenseal~=0.3.15
pillow~=11.0.0
tqdm~=4.66.6